+++ src/assets/index.ts
@@ -1,4 +1,5 @@
 export * from "./asset";
 export * from "./bitmapFont";
-export * from "./fonts";
+export * from "./font";
 export * from "./shader";
+export * from "./sprite";
