+++ src/components/draw/sprite.ts
@@ -1,12 +1,12 @@
 // TODO: accept canvas
 
+import type { Asset } from "../../assets";
 import { resolveSprite } from "../../assets/sprite";
 import { getRenderProps } from "../../game/utils";
 import { drawTexture, type Texture } from "../../gfx";
 import { k } from "../../kaboom";
 import { Quad, quad, type Rect, Vec2, vec2 } from "../../math";
 import type {
-    Asset,
     Comp,
     GameObj,
     KaboomCtx,
