+++ src/components/physics/area.ts
@@ -11,13 +11,13 @@ import type {
     GameObj,
     KaboomCtx,
     MouseButton,
-    PosComp,
     Shape,
     Tag,
 } from "../../types";
 import type { KEventController } from "../../utils/";
 import type { AnchorComp } from "../transform/anchor";
 import type { FixedComp } from "../transform/fixed";
+import type { PosComp } from "../transform/pos";
 
 /**
  * The {@link area `area()`} component.
