+++ src/components/transform/scale.ts
@@ -1,5 +1,5 @@
 import { type Vec2, vec2, type Vec2Args } from "../../math";
-import type { Comp, KaboomCtx } from "../../types";
+import type { Comp } from "../../types";
 
 /**
  * The {@link scale `scale()`} component.
