+++ src/gfx/draw/drawRaw.ts
@@ -1,8 +1,8 @@
+import { Asset } from "../../assets";
 import { resolveShader } from "../../assets/shader";
 import { game, gfx } from "../../kaboom";
 import { screen2ndc } from "../../math/various";
 import type { RenderProps, Uniform, Vertex } from "../../types";
-import { Asset } from "../assets";
 import type { Texture } from "../gfx";
 
 export function drawRaw(
