+++ src/gfx/draw/index.ts
@@ -1,20 +1,20 @@
-export * from './drawBezier';
-export * from './drawCircle';
-export * from './drawCurve';
-
-export * from './drawEllipse';
-export * from './drawFormattedText';
-export * from './drawInspectText';
-export * from './drawLine';
-export * from './drawMasked';
-export * from './drawPolygon';
-export * from './drawRaw';
-export * from './drawRect';
-export * from './drawSprite';
-export * from './drawStenciled';
-export * from './drawSubstracted';
-export * from './drawText';
-export * from './drawTexture';
-export * from './drawTriangle';
-export * from './drawUVQuad';
-export * from './drawUnscaled';
+export * from "./drawBezier";
+export * from "./drawCircle";
+export * from "./drawCurve";
+export * from "./drawDebug";
+export * from "./drawEllipse";
+export * from "./drawFormattedText";
+export * from "./drawInspectText";
+export * from "./drawLine";
+export * from "./drawMasked";
+export * from "./drawPolygon";
+export * from "./drawRaw";
+export * from "./drawRect";
+export * from "./drawSprite";
+export * from "./drawStenciled";
+export * from "./drawSubstracted";
+export * from "./drawText";
+export * from "./drawTexture";
+export * from "./drawTriangle";
+export * from "./drawUnscaled";
+export * from "./drawUVQuad";
