+++ src/gfx/index.ts
@@ -1,6 +1,4 @@
 export * from "./anchor";
-export * from "./assets";
-export * from "./assets";
 export * from "./draw";
 export * from "./formatText";
 export * from "./gfx";
